myproj/requirements.txt
Django==3.2.6
gunicorn==20.1.0
mysqlclient==2.1.1
pandas==1.4.3
django-debug-toolbar==3.2.4
djangorestframework==3.14.0
PyMySQL==1.0.2

